backend/requirements.txt
google-cloud-aiplatform==1.43.0
langchain==0.1.10
langchain-community==0.0.25
langchain-google-vertexai
typing_extensions==4.9.0
unstructured==0.12.4
pdf2image==1.17.0
pytesseract==0.3.10
pdfminer.six==20221105
pillow-heif==0.15.0
opencv-python==4.9.0.80
unstructured-inference==0.7.24
pikepdf==8.13.0
pypdf==4.0.1
tensorflow_hub==0.16.1
tensorflow_text
tensorboard==2.16
protobuf==3.20.3
pipdeptree 
ipykernel
pydantic
fastapi
uvicorn[standard]
